chore: move write buffer into its own crate
